+++ src/partner-shop-application-api/tests/get_all.rs
@@ -8,6 +8,8 @@ use partner_shop_application::{
     },
 };
 use partner_shop_application_api::handler;
+use shop::dynamodb::repository::ShopDynamoDbRepositoryImpl;
+use shop::service::get_service::GetShopServiceImpl;
 use test_api::*;
 use user::dynamodb::repository::UserDynamoDbRepositoryImpl;
 use user::service::user_service::UserServiceImpl;
@@ -26,8 +28,11 @@ async fn should_200_respond_empty_list_when_no_applications_exist() {
     sfn_adapter
         .expect_start_execution()
         .returning(|_, _| Box::pin(async { Ok("foo".into()) }));
+    let shop_repository = ShopDynamoDbRepositoryImpl::new(get_dynamodb_client().await, "table_1");
+    let shop_service = GetShopServiceImpl::new(&shop_repository);
     let service = PartnerShopApplicationServiceImpl::new(
         &repository,
+        &shop_service,
         &sfn_adapter,
         "arn:aws:states:us-east-1:123456789:stateMachine:test",
     );
@@ -59,8 +64,11 @@ async fn should_200_respond_applications_when_they_exist() {
     sfn_adapter
         .expect_start_execution()
         .returning(|_, _| Box::pin(async { Ok("foo".into()) }));
+    let shop_repository = ShopDynamoDbRepositoryImpl::new(get_dynamodb_client().await, "table_1");
+    let shop_service = GetShopServiceImpl::new(&shop_repository);
     let service = PartnerShopApplicationServiceImpl::new(
         &repository,
+        &shop_service,
         &sfn_adapter,
         "arn:aws:states:us-east-1:123456789:stateMachine:test",
     );