+++ src/partner-shop-application-api/tests/patch.rs
@@ -13,6 +13,8 @@ use partner_shop_application::{
     },
 };
 use partner_shop_application_api::handler;
+use shop::dynamodb::repository::ShopDynamoDbRepositoryImpl;
+use shop::service::get_service::GetShopServiceImpl;
 use test_api::*;
 use user::dynamodb::repository::UserDynamoDbRepositoryImpl;
 use user::service::user_service::UserServiceImpl;
@@ -31,8 +33,11 @@ async fn should_200_when_updating_application() {
     sfn_adapter
         .expect_start_execution()
         .returning(|_, _| Box::pin(async { Ok("foo".into()) }));
+    let shop_repository = ShopDynamoDbRepositoryImpl::new(get_dynamodb_client().await, "table_1");
+    let shop_service = GetShopServiceImpl::new(&shop_repository);
     let service = PartnerShopApplicationServiceImpl::new(
         &repository,
+        &shop_service,
         &sfn_adapter,
         "arn:aws:states:us-east-1:123456789:stateMachine:test",
     );